refactor: changing import
