+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import { MaxWidthWrapper } from "./components/maxWidthWrapper/MaxWidthWrapper";
+import { MaxWidthWrapper } from "./components/ui/maxWidthWrapper/MaxWidthWrapper";
 import { FeatureSection } from "./pages/home/featureSection/FeatureSection";
 import { HeroSection } from "./pages/home/heroSection/HeroSection";
 import { PropositionsSection } from "./pages/home/propositionSection/PropositionSection";
